Require transaction support
The existing transaction-detection logic is faulty, so
it needed to be replaced.  But I don't think it's correct
to run against a MongoDB server that doesn't support transactions,
so best to just assume they are supported and let the server
error if they are not.

A simple way to launch a replica set (which supports transactions)
is with mlaunch (https://rueckstiess.github.io/mtools/mlaunch.html).
A single node replica set should be sufficient.