Customize parameter marker so that it creates valid JSON
Using '?' for parameter markers is problematic because it creates unparseable JSON,
and required the JDBC driver to search for '?' using brittle code that doesn't
currently account for the use of '?' in other places, e.g. within a constant string
(I'm not sure how JDBC drivers handle this, actually).

The other problem with the current approach is that it requires all parameters to be
serialized to a JSON string, only to be subsequently parsed into a BsonValue.

This patch proposes an alternative. Instead of using '?', take advantage of the fact
that there is a long-deprecated BSON type called Undefined, whose extended JSON
representation is "{$undefined: true}".  By using that String as the custom parameter
marker, we can parse the JSON string without any pre-processing, and then recursively
replace all instances of BsonUndefined with the BsonValue corresponding to the parameter
with the same index.

This approach requires that hibernate.dialect.native_param_markers be set to true
(Postgres also does this).

One problem is that it broke the native query test.  I worked around this by removing
the parameter from the native query. We will probably have to figure out how to get
that working again.