WIP of introducing a Mongo AST based on the C# driver
I integrated it into queries so far, but not mutations, but I think it's
all fairly straightforward to do.

The main issue is state management while building the AST.  The most
natural thing would be to return AST nodes or their properties from various
methods, but many of those methods are defined by existing interfaces to
return void.  So I ended up adding a bunch of fields to the translator class
in order to transfer state from callee back to caller, and it seems
inelegant and error-prone.